Add comments & second half of gameReset
Added the second half of gameReset which takes the size input by the user and divides it
evenly. This however, needs refactor and it only temp for testing. for instance, leaving
the code this way opens the door wide open for errors in uneven numbers and adding floats
into the mix.