Add while condition for input error & add parseInt wrapper to size
Added a while loop that executes while the input size is out of bounds, or user inputs
non integer values. In order to determine if an integer was input, the size needed conversion
so a parseInt() wrapper was put in place. Now if the user inputs say a string, it will trigger
a false flag in the while logic, forcing the loop to re-occur until a integer in bounds is placed
in