Remove while loop and replace with if shorthand
Replaced the while loop with a simple if statement that displays the error message if
the flags are fired, and throws the gameReset() function if they are not. The reasoning
here being they can't really access the current game should they wish to continue either
by accidently clicking reset or realizing the desired game they want, is not capable within
the bounds.