Add rainbow script code
Added the code which takes the class rainbow in html and checks if the button has been clicked
the num variable is set to zero each call of the draw function. if the rainbow button has been
clicked the function runs a conditional determining if the button click is even or odd. if even
the rainbow variable is set to true, if odd the rainbow value returns to false. This allows rainbow
to be treated more like a switch than a button. within the mouseenter event, there is another
conditional. If rainbow is true a random color will be generated and amended to target. Else
the target defaults to black.